v4l2bufferpool: Change the buffer number to be queued into V4L2
Apparently the continuous frame capturing mode of VIN has to
have at least 3 buffers to be queued into the driver in order to
output frames and the GStreamer pipeline occupies up to 2 buffers
, which correspond to the front and back buffers in the videosink.
The total number to circulate the V4L2 buffers in the pipeline is 5,
so this patch sets that as the buffer pool parameter.

Also this change suppresses the buffer copying that occurs when
running out of the remaining buffers in the queue.