v4l2src/v4l2bufferpool: Set the required number of the pool buffers in decide_allocation istead of buffer_pool_new
The number of the buffers in the pool should be determined in
decide_allocation(), so this patch moves the setting of the minimun
and maximum numbers of the buffers to use the contiguous capturing mode
from buffer_pool_new() to decide_allocation().

The setting in buffer_pool_new() has reverted.